review changes - link the symbols in documentation for objc * swift changes will be done later
review changes

+++ Objective-C/CBLReplicatorTypes.h
@@ -42,7 +42,7 @@ typedef struct {
 } CBLReplicatorProgress;
 
 /** Replicator type. */
-typedef NS_ENUM(uint8_t, CBLReplicatorType) {
+typedef NS_ENUM(NSUInteger, CBLReplicatorType) {
     kCBLReplicatorTypePushAndPull = 0,              ///< Bidirectional; both push and pull
     kCBLReplicatorTypePush,                         ///< Pushing changes to the target
     kCBLReplicatorTypePull                          ///< Pulling changes from the target
